removed edge case for routines access
