+++ internal/corazawaf/transaction.go
@@ -105,7 +105,7 @@ type Transaction struct {
 	// We must reuse it in the future
 	Capture bool
 
-	// Contains duration in useconds per phase
+	// Contains duration in microseconds per phase
 	stopWatches map[types.RulePhase]int64
 
 	// Contains a WAF instance for the current transaction
@@ -408,12 +408,10 @@ func (tx *Transaction) CaptureField(index int, value string) {
 // this function is used to control which variables are reset after a new rule is evaluated
 func (tx *Transaction) resetCaptures() {
 	tx.debugLogger.Debug().
-		Msg("Reseting captured variables")
+		Msg("Resetting captured variables")
 	// We reset capture 0-9
 	ctx := tx.variables.tx
-	// RUNE 48 = 0
-	// RUNE 57 = 9
-	for i := rune(48); i <= 57; i++ {
+	for i := rune('0'); i <= '9'; i++ {
 		ctx.SetIndex(string(i), 0, "")
 	}
 }
@@ -848,6 +846,51 @@ func setAndReturnBodyLimitInterruption(tx *Transaction, status int) (*types.Inte
 	return tx.interruption, 0, nil
 }
 
+// bodyLimitReached handles the case where the body buffer has already reached its limit.
+// Returns (interruption, shouldReturn). When shouldReturn is true, the caller should
+// return the interruption (which may be nil for ProcessPartial).
+func (tx *Transaction) bodyLimitReached(bufLen, limit int64, limitAction types.BodyLimitAction) (*types.Interruption, bool) {
+	if limit == bufLen {
+		if limitAction == types.BodyLimitActionReject {
+			return tx.interruption, true
+		}
+		if limitAction == types.BodyLimitActionProcessPartial {
+			return nil, true
+		}
+	}
+	return nil, false
+}
+
+// checkBodyOverflow checks whether writing writingBytes to a body buffer of bufLen would exceed limit.
+// It performs an integer overflow check and returns:
+//   - clampedBytes: the number of bytes safe to write
+//   - runProcess: whether the body processor should run (limit reached with ProcessPartial)
+//   - interruption: non-nil if the transaction should be interrupted
+//   - err: non-nil on integer overflow
+func (tx *Transaction) checkBodyOverflow(
+	bufLen, writingBytes, limit int64,
+	limitAction types.BodyLimitAction,
+	rejectStatus int,
+	dataErrorVar *collections.Single,
+) (clampedBytes int64, runProcess bool, interruption *types.Interruption, err error) {
+	// Overflow check: MaxInt64 is not a realistic payload size
+	if bufLen >= (math.MaxInt64 - writingBytes) {
+		return 0, false, nil, errors.New("overflow reached while writing body")
+	}
+
+	if bufLen+writingBytes >= limit {
+		dataErrorVar.Set("1")
+		if limitAction == types.BodyLimitActionReject {
+			interruption, _, _ := setAndReturnBodyLimitInterruption(tx, rejectStatus)
+			return 0, false, interruption, nil
+		}
+		if limitAction == types.BodyLimitActionProcessPartial {
+			return limit - bufLen, true, nil, nil
+		}
+	}
+	return writingBytes, false, nil, nil
+}
+
 // WriteRequestBody writes bytes from a slice of bytes into the request body,
 // it returns an interruption if the writing bytes go beyond the request body limit.
 // It won't copy the bytes if the body access isn't accessible.
@@ -860,40 +903,19 @@ func (tx *Transaction) WriteRequestBody(b []byte) (*types.Interruption, int, err
 		return nil, 0, nil
 	}
 
-	if tx.RequestBodyLimit == tx.requestBodyBuffer.length {
-		// tx.RequestBodyLimit will never be zero so if this happened, we have an
-		// interruption (that has been previously raised, but ignored by the connector) for sure.
-		if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionReject {
-			return tx.interruption, 0, nil
-		}
-
-		if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionProcessPartial {
-			return nil, 0, nil
-		}
+	if it, shouldReturn := tx.bodyLimitReached(tx.requestBodyBuffer.length, tx.RequestBodyLimit, tx.WAF.RequestBodyLimitAction); shouldReturn {
+		return it, 0, nil
 	}
 
-	var (
-		writingBytes          = int64(len(b))
-		runProcessRequestBody = false
+	writingBytes, runProcess, it, err := tx.checkBodyOverflow(
+		tx.requestBodyBuffer.length, int64(len(b)), tx.RequestBodyLimit,
+		tx.WAF.RequestBodyLimitAction, 413, tx.variables.inboundDataError,
 	)
-	// Overflow check
-	if tx.requestBodyBuffer.length >= (math.MaxInt64 - writingBytes) {
-		// Overflow, failing. MaxInt64 is not a realistic payload size. Furthermore, it has been tested that
-		// bytes.Buffer does not work with this kind of sizes. See comments in BodyBuffer Write(data []byte)
-		return nil, 0, errors.New("overflow reached while writing request body")
+	if err != nil {
+		return nil, 0, err
 	}
-
-	if tx.requestBodyBuffer.length+writingBytes >= tx.RequestBodyLimit {
-		tx.variables.inboundDataError.Set("1")
-		if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionReject {
-			// We interrupt this transaction in case RequestBodyLimitAction is Reject
-			return setAndReturnBodyLimitInterruption(tx, 413)
-		}
-
-		if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionProcessPartial {
-			writingBytes = tx.RequestBodyLimit - tx.requestBodyBuffer.length
-			runProcessRequestBody = true
-		}
+	if it != nil {
+		return it, 0, nil
 	}
 
 	w, err := tx.requestBodyBuffer.Write(b[:writingBytes])
@@ -901,7 +923,7 @@ func (tx *Transaction) WriteRequestBody(b []byte) (*types.Interruption, int, err
 		return nil, 0, err
 	}
 
-	if runProcessRequestBody {
+	if runProcess {
 		tx.debugLogger.Warn().Msg("Processing request body whose size reached the configured limit (Action ProcessPartial)")
 		_, err = tx.ProcessRequestBody()
 	}
@@ -925,40 +947,26 @@ func (tx *Transaction) ReadRequestBodyFrom(r io.Reader) (*types.Interruption, in
 		return nil, 0, nil
 	}
 
-	if tx.RequestBodyLimit == tx.requestBodyBuffer.length {
-		// tx.RequestBodyLimit will never be zero so if this happened, we have an
-		// interruption (that has been previously raised, but ignored by the connector) for sure.
-		if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionReject {
-			return tx.interruption, 0, nil
-		}
-
-		if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionProcessPartial {
-			return nil, 0, nil
-		}
+	if it, shouldReturn := tx.bodyLimitReached(tx.requestBodyBuffer.length, tx.RequestBodyLimit, tx.WAF.RequestBodyLimitAction); shouldReturn {
+		return it, 0, nil
 	}
 
 	var (
-		writingBytes          int64
-		runProcessRequestBody = false
+		writingBytes int64
+		runProcess   = false
 	)
 	if l, ok := r.(ByteLenger); ok {
-		writingBytes = int64(l.Len())
-		// Overflow check
-		if tx.requestBodyBuffer.length >= (math.MaxInt64 - writingBytes) {
-			// Overflow, failing. MaxInt64 is not a realistic payload size. Furthermore, it has been tested that
-			// bytes.Buffer does not work with this kind of sizes. See comments in BodyBuffer Write(data []byte)
-			return nil, 0, errors.New("overflow reached while writing request body")
+		var it *types.Interruption
+		var err error
+		writingBytes, runProcess, it, err = tx.checkBodyOverflow(
+			tx.requestBodyBuffer.length, int64(l.Len()), tx.RequestBodyLimit,
+			tx.WAF.RequestBodyLimitAction, 413, tx.variables.inboundDataError,
+		)
+		if err != nil {
+			return nil, 0, err
 		}
-		if tx.requestBodyBuffer.length+writingBytes >= tx.RequestBodyLimit {
-			tx.variables.inboundDataError.Set("1")
-			if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionReject {
-				return setAndReturnBodyLimitInterruption(tx, 413)
-			}
-
-			if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionProcessPartial {
-				writingBytes = tx.RequestBodyLimit - tx.requestBodyBuffer.length
-				runProcessRequestBody = true
-			}
+		if it != nil {
+			return it, 0, nil
 		}
 	} else {
 		writingBytes = tx.RequestBodyLimit - tx.requestBodyBuffer.length
@@ -976,12 +984,12 @@ func (tx *Transaction) ReadRequestBodyFrom(r io.Reader) (*types.Interruption, in
 		}
 
 		if tx.WAF.RequestBodyLimitAction == types.BodyLimitActionProcessPartial {
-			runProcessRequestBody = true
+			runProcess = true
 		}
 	}
 
 	err = nil
-	if runProcessRequestBody {
+	if runProcess {
 		tx.debugLogger.Warn().Msg("Processing request body whose size reached the configured limit (Action ProcessPartial)")
 		_, err = tx.ProcessRequestBody()
 	}
@@ -1024,56 +1032,61 @@ func (tx *Transaction) ProcessRequestBody() (*types.Interruption, error) {
 		tx.WAF.Rules.Eval(types.PhaseRequestBody, tx)
 		return tx.interruption, nil
 	}
-	mime := ""
-	if m := tx.variables.requestHeaders.Get("content-type"); len(m) > 0 {
-		mime = m[0]
-	}
 
 	reader, err := tx.requestBodyBuffer.Reader()
 	if err != nil {
 		return nil, err
 	}
 
-	rbp := tx.variables.reqbodyProcessor.Get()
+	if err := tx.selectAndRunRequestBodyProcessor(reader); err != nil {
+		return nil, err
+	}
+
+	tx.WAF.Rules.Eval(types.PhaseRequestBody, tx)
+	return tx.interruption, nil
+}
+
+// selectAndRunRequestBodyProcessor resolves the body processor name, looks up the processor,
+// and runs it against the given reader. It returns an error only for unexpected failures.
+func (tx *Transaction) selectAndRunRequestBodyProcessor(reader io.Reader) error {
+	mime := ""
+	if m := tx.variables.requestHeaders.Get("content-type"); len(m) > 0 {
+		mime = m[0]
+	}
+
+	bodyProcessorName := tx.variables.reqbodyProcessor.Get()
 
 	// Default variables.ReqbodyProcessor values
 	// XML and JSON must be forced with ctl:requestBodyProcessor=JSON
 	if tx.ForceRequestBodyVariable {
-		// We force URLENCODED if mime is x-www... or we have an empty RBP and ForceRequestBodyVariable
-		if rbp == "" {
-			rbp = "URLENCODED"
+		// We force URLENCODED if mime is x-www... or we have an empty body processor and ForceRequestBodyVariable
+		if bodyProcessorName == "" {
+			bodyProcessorName = "URLENCODED"
 		}
-		tx.variables.reqbodyProcessor.Set(rbp)
+		tx.variables.reqbodyProcessor.Set(bodyProcessorName)
 	}
-	rbp = strings.ToLower(rbp)
-	if rbp == "" {
-		// so there is no bodyprocessor, we don't want to generate an error
-		tx.WAF.Rules.Eval(types.PhaseRequestBody, tx)
-		return tx.interruption, nil
+	bodyProcessorName = strings.ToLower(bodyProcessorName)
+	if bodyProcessorName == "" {
+		return nil
 	}
-	bodyprocessor, err := bodyprocessors.GetBodyProcessor(rbp)
+	bp, err := bodyprocessors.GetBodyProcessor(bodyProcessorName)
 	if err != nil {
 		tx.generateRequestBodyError(errors.New("invalid body processor"))
-		tx.WAF.Rules.Eval(types.PhaseRequestBody, tx)
-		return tx.interruption, nil
+		return nil
 	}
 
 	tx.debugLogger.Debug().
-		Str("body_processor", rbp).
+		Str("body_processor", bodyProcessorName).
 		Msg("Attempting to process request body")
 
-	if err := bodyprocessor.ProcessRequest(reader, tx.Variables(), plugintypes.BodyProcessorOptions{
+	if err := bp.ProcessRequest(reader, tx.Variables(), plugintypes.BodyProcessorOptions{
 		Mime:        mime,
 		StoragePath: tx.WAF.UploadDir,
 	}); err != nil {
 		tx.debugLogger.Error().Err(err).Msg("Failed to process request body")
 		tx.generateRequestBodyError(err)
-		tx.WAF.Rules.Eval(types.PhaseRequestBody, tx)
-		return tx.interruption, nil
 	}
-
-	tx.WAF.Rules.Eval(types.PhaseRequestBody, tx)
-	return tx.interruption, nil
+	return nil
 }
 
 // ProcessResponseHeaders performs the analysis on the response headers.
@@ -1098,8 +1111,8 @@ func (tx *Transaction) ProcessResponseHeaders(code int, proto string) *types.Int
 		return tx.interruption
 	}
 
-	c := strconv.Itoa(code)
-	tx.variables.responseStatus.Set(c)
+	statusCode := strconv.Itoa(code)
+	tx.variables.responseStatus.Set(statusCode)
 	tx.variables.responseProtocol.Set(proto)
 
 	tx.WAF.Rules.Eval(types.PhaseResponseHeaders, tx)
@@ -1133,40 +1146,27 @@ func (tx *Transaction) WriteResponseBody(b []byte) (*types.Interruption, int, er
 		return nil, 0, nil
 	}
 
-	if tx.ResponseBodyLimit == tx.responseBodyBuffer.length {
-		// tx.ResponseBodyLimit will never be zero so if this happened, we have an
-		// interruption for sure.
-		if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionReject {
-			return tx.interruption, 0, nil
-		}
-
-		if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionProcessPartial {
-			return nil, 0, nil
-		}
+	if it, shouldReturn := tx.bodyLimitReached(tx.responseBodyBuffer.length, tx.ResponseBodyLimit, tx.WAF.ResponseBodyLimitAction); shouldReturn {
+		return it, 0, nil
 	}
 
-	var (
-		writingBytes           = int64(len(b))
-		runProcessResponseBody = false
+	writingBytes, runProcess, it, err := tx.checkBodyOverflow(
+		tx.responseBodyBuffer.length, int64(len(b)), tx.ResponseBodyLimit,
+		tx.WAF.ResponseBodyLimitAction, 500, tx.variables.outboundDataError,
 	)
-	if tx.responseBodyBuffer.length+writingBytes >= tx.ResponseBodyLimit {
-		tx.variables.outboundDataError.Set("1")
-		if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionReject {
-			// We interrupt this transaction in case ResponseBodyLimitAction is Reject
-			return setAndReturnBodyLimitInterruption(tx, 500)
-		}
-
-		if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionProcessPartial {
-			writingBytes = tx.ResponseBodyLimit - tx.responseBodyBuffer.length
-			runProcessResponseBody = true
-		}
+	if err != nil {
+		return nil, 0, err
 	}
+	if it != nil {
+		return it, 0, nil
+	}
+
 	w, err := tx.responseBodyBuffer.Write(b[:writingBytes])
 	if err != nil {
 		return nil, 0, err
 	}
 
-	if runProcessResponseBody {
+	if runProcess {
 		_, err = tx.ProcessResponseBody()
 	}
 	return tx.interruption, int(w), err
@@ -1184,32 +1184,26 @@ func (tx *Transaction) ReadResponseBodyFrom(r io.Reader) (*types.Interruption, i
 		return nil, 0, nil
 	}
 
-	if tx.ResponseBodyLimit == tx.responseBodyBuffer.length {
-		if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionReject {
-			return tx.interruption, 0, nil
-		}
-
-		if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionProcessPartial {
-			return nil, 0, nil
-		}
+	if it, shouldReturn := tx.bodyLimitReached(tx.responseBodyBuffer.length, tx.ResponseBodyLimit, tx.WAF.ResponseBodyLimitAction); shouldReturn {
+		return it, 0, nil
 	}
 
 	var (
-		writingBytes           int64
-		runProcessResponseBody = false
+		writingBytes int64
+		runProcess   = false
 	)
 	if l, ok := r.(ByteLenger); ok {
-		writingBytes = int64(l.Len())
-		if tx.responseBodyBuffer.length+writingBytes >= tx.ResponseBodyLimit {
-			tx.variables.outboundDataError.Set("1")
-			if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionReject {
-				return setAndReturnBodyLimitInterruption(tx, 500)
-			}
-
-			if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionProcessPartial {
-				writingBytes = tx.ResponseBodyLimit - tx.responseBodyBuffer.length
-				runProcessResponseBody = true
-			}
+		var it *types.Interruption
+		var err error
+		writingBytes, runProcess, it, err = tx.checkBodyOverflow(
+			tx.responseBodyBuffer.length, int64(l.Len()), tx.ResponseBodyLimit,
+			tx.WAF.ResponseBodyLimitAction, 500, tx.variables.outboundDataError,
+		)
+		if err != nil {
+			return nil, 0, err
+		}
+		if it != nil {
+			return it, 0, nil
 		}
 	} else {
 		writingBytes = tx.ResponseBodyLimit - tx.responseBodyBuffer.length
@@ -1227,18 +1221,18 @@ func (tx *Transaction) ReadResponseBodyFrom(r io.Reader) (*types.Interruption, i
 		}
 
 		if tx.WAF.ResponseBodyLimitAction == types.BodyLimitActionProcessPartial {
-			runProcessResponseBody = true
+			runProcess = true
 		}
 	}
 
 	err = nil
-	if runProcessResponseBody {
+	if runProcess {
 		_, err = tx.ProcessResponseBody()
 	}
 	return tx.interruption, int(w), err
 }
 
-// ProcessResponseBody Perform the analysis of the the response body (if any)
+// ProcessResponseBody Perform the analysis of the response body (if any)
 //
 // It is recommended to call this method even if it is not expected to have a body.
 // It permits to execute rules belonging to request body phase, but not necessarily
@@ -1285,12 +1279,22 @@ func (tx *Transaction) ProcessResponseBody() (*types.Interruption, error) {
 		return tx.interruption, err
 	}
 
+	if err := tx.selectAndRunResponseBodyProcessor(reader); err != nil {
+		return tx.interruption, err
+	}
+
+	tx.WAF.Rules.Eval(types.PhaseResponseBody, tx)
+	return tx.interruption, nil
+}
+
+// selectAndRunResponseBodyProcessor resolves the response body processor and runs it,
+// or falls back to reading the response body into variables.
+func (tx *Transaction) selectAndRunResponseBodyProcessor(reader io.Reader) error {
 	if bp := tx.variables.resBodyProcessor.Get(); bp != "" {
 		b, err := bodyprocessors.GetBodyProcessor(bp)
 		if err != nil {
 			tx.generateResponseBodyError(errors.New("invalid body processor"))
-			tx.WAF.Rules.Eval(types.PhaseResponseBody, tx)
-			return tx.interruption, err
+			return err
 		}
 
 		tx.debugLogger.Debug().Str("body_processor", bp).Msg("Attempting to process response body")
@@ -1299,17 +1303,17 @@ func (tx *Transaction) ProcessResponseBody() (*types.Interruption, error) {
 			tx.debugLogger.Error().Err(err).Msg("Failed to process response body")
 			tx.generateResponseBodyError(err)
 		}
-	} else {
-		buf := new(strings.Builder)
-		length, err := io.Copy(buf, reader)
-		if err != nil {
-			return tx.interruption, err
-		}
-		tx.variables.responseContentLength.Set(strconv.FormatInt(length, 10))
-		tx.variables.responseBody.Set(buf.String())
+		return nil
 	}
-	tx.WAF.Rules.Eval(types.PhaseResponseBody, tx)
-	return tx.interruption, nil
+
+	buf := new(strings.Builder)
+	length, err := io.Copy(buf, reader)
+	if err != nil {
+		return err
+	}
+	tx.variables.responseContentLength.Set(strconv.FormatInt(length, 10))
+	tx.variables.responseBody.Set(buf.String())
+	return nil
 }
 
 // ProcessLogging logs all information relative to this transaction.
@@ -1401,8 +1405,14 @@ func (tx *Transaction) AuditLog() *auditlog.Log {
 	al := &auditlog.Log{}
 	al.Parts_ = tx.AuditLogParts
 
-	clientPort, _ := strconv.Atoi(tx.variables.remotePort.Get())
-	hostPort, _ := strconv.Atoi(tx.variables.serverPort.Get())
+	clientPort, err := strconv.Atoi(tx.variables.remotePort.Get())
+	if err != nil {
+		tx.debugLogger.Debug().Err(err).Msg("Error converting client port to integer")
+	}
+	hostPort, err := strconv.Atoi(tx.variables.serverPort.Get())
+	if err != nil {
+		tx.debugLogger.Debug().Err(err).Msg("Error converting host port to integer")
+	}
 
 	// Convert the transaction fullRequestLength to Int32
 	requestLength, err := strconv.ParseInt(tx.variables.fullRequestLength.Get(), 10, 32)
@@ -1436,123 +1446,149 @@ func (tx *Transaction) AuditLog() *auditlog.Log {
 		IsInterrupted_: tx.IsInterrupted(),
 	}
 
-	var auditLogPartAuditLogTrailerSet, auditLogPartRulesMatchedSet bool
+	var includeErrorLog, rulesMatchedSet bool
 	for _, part := range tx.AuditLogParts {
 		switch part {
 		case types.AuditLogPartRequestHeaders:
-			al.Transaction_.Request_.Headers_ = tx.variables.requestHeaders.Data()
+			tx.auditLogPartRequestHeaders(al)
 		case types.AuditLogPartRequestBody:
-			reader, err := tx.requestBodyBuffer.Reader()
-			if err == nil {
-				content, err := io.ReadAll(reader)
-				if err == nil {
-					al.Transaction_.Request_.Body_ = string(content)
-				}
-			}
-
-			/*
-			* TODO:
-			* This part is a replacement for part C. It will log the same data as C in
-			* all cases except when multipart/form-data encoding in used. In this case,
-			* it will log a fake application/x-www-form-urlencoded body that contains
-			* the information about parameters but not about the files. This is handy
-			* if you don’t want to have (often large) files stored in your audit logs.
-			 */
-			// upload data
-			var files []plugintypes.AuditLogTransactionRequestFiles
-			al.Transaction_.Request_.Files_ = nil
-			for _, file := range tx.variables.files.Get("") {
-				var size int64
-				if fs := tx.variables.filesSizes.Get(file); len(fs) > 0 {
-					size, _ = strconv.ParseInt(fs[0], 10, 64)
-					// we ignore the error as it defaults to 0
-				}
-				ext := filepath.Ext(file)
-				at := auditlog.TransactionRequestFiles{
-					Size_: size,
-					Name_: file,
-					Mime_: mime.TypeByExtension(ext),
-				}
-				files = append(files, at)
-			}
-			al.Transaction_.Request_.Files_ = files
+			tx.auditLogPartRequestBody(al)
 		case types.AuditLogPartIntermediaryResponseBody:
-			if al.Transaction_.Response_ == nil {
-				al.Transaction_.Response_ = &auditlog.TransactionResponse{}
-			}
-			al.Transaction_.Response_.Body_ = tx.variables.responseBody.Get()
+			tx.auditLogPartResponseBody(al)
 		case types.AuditLogPartResponseHeaders:
-			if al.Transaction_.Response_ == nil {
-				al.Transaction_.Response_ = &auditlog.TransactionResponse{}
-			}
-			status, _ := strconv.Atoi(tx.variables.responseStatus.Get())
-			al.Transaction_.Response_.Status_ = status
-			al.Transaction_.Response_.Headers_ = tx.variables.responseHeaders.Data()
+			tx.auditLogPartResponseHeaders(al)
 		case types.AuditLogPartAuditLogTrailer:
-			auditLogPartAuditLogTrailerSet = true
-			al.Transaction_.Producer_ = &auditlog.TransactionProducer{
-				Connector_:  tx.WAF.ProducerConnector,
-				Version_:    tx.WAF.ProducerConnectorVersion,
-				Server_:     "",
-				RuleEngine_: tx.RuleEngine.String(),
-				Stopwatch_:  tx.GetStopWatch(),
-				Rulesets_:   tx.WAF.ComponentNames,
-			}
+			includeErrorLog = true
+			tx.auditLogPartProducer(al)
 		case types.AuditLogPartRulesMatched:
-			auditLogPartRulesMatchedSet = true
-			for _, mr := range tx.matchedRules {
-				// Log action is required to log a matched rule on both error log and audit log
-				// An assertion has to be done to check if the MatchedRule implements the Log() function before calling Log()
-				// It is performed to avoid breaking the Coraza v3.* API adding a Log() method to the MatchedRule interface
-				mrWithlog, ok := mr.(*corazarules.MatchedRule)
-				if ok && mrWithlog.Log() {
-					r := mr.Rule()
-					for _, matchData := range mr.MatchedDatas() {
-						newAlEntry := auditlog.Message{
-							Actionset_: strings.Join(tx.WAF.ComponentNames, " "),
-							Message_:   matchData.Message(),
-							Data_: &auditlog.MessageData{
-								File_:     mr.Rule().File(),
-								Line_:     mr.Rule().Line(),
-								ID_:       r.ID(),
-								Rev_:      r.Revision(),
-								Msg_:      matchData.Message(),
-								Data_:     matchData.Data(),
-								Severity_: r.Severity(),
-								Ver_:      r.Version(),
-								Maturity_: r.Maturity(),
-								Accuracy_: r.Accuracy(),
-								Tags_:     r.Tags(),
-								Raw_:      r.Raw(),
-							},
-						}
-						// If AuditLogPartAuditLogTrailer (H) is set, we expect to log the error messages emitted by the rules
-						// in the audit log
-						if auditLogPartAuditLogTrailerSet {
-							newAlEntry.ErrorMessage_ = mr.ErrorLog()
-						}
-						al.Messages_ = append(al.Messages_, newAlEntry)
-					}
-				}
-			}
+			rulesMatchedSet = true
+			tx.auditLogPartMatchedRules(al, includeErrorLog)
 		}
 	}
 
 	// If AuditLogPartRulesMatched (K) is not set, but AuditLogPartAuditLogTrailer (H) is set, we still expect to
 	// log the error messages emitted by the rules (if the rule has Log set to true)
-	if !auditLogPartRulesMatchedSet && auditLogPartAuditLogTrailerSet {
-		for _, mr := range tx.matchedRules {
-			mrWithlog, ok := mr.(*corazarules.MatchedRule)
-			if ok && mrWithlog.Log() {
+	if !rulesMatchedSet && includeErrorLog {
+		for _, matchedRule := range tx.matchedRules {
+			loggableRule, ok := matchedRule.(*corazarules.MatchedRule)
+			if ok && loggableRule.Log() {
 				al.Messages_ = append(al.Messages_, auditlog.Message{
-					ErrorMessage_: mr.ErrorLog(),
+					ErrorMessage_: matchedRule.ErrorLog(),
 				})
 			}
 		}
-
 	}
 
 	return al
+}
+
+func (tx *Transaction) auditLogPartRequestHeaders(al *auditlog.Log) {
+	al.Transaction_.Request_.Headers_ = tx.variables.requestHeaders.Data()
+}
+
+func (tx *Transaction) auditLogPartRequestBody(al *auditlog.Log) {
+	reader, err := tx.requestBodyBuffer.Reader()
+	if err == nil {
+		content, err := io.ReadAll(reader)
+		if err == nil {
+			al.Transaction_.Request_.Body_ = string(content)
+		}
+	}
+
+	/*
+	* TODO:
+	* This part is a replacement for part C. It will log the same data as C in
+	* all cases except when multipart/form-data encoding in used. In this case,
+	* it will log a fake application/x-www-form-urlencoded body that contains
+	* the information about parameters but not about the files. This is handy
+	* if you don’t want to have (often large) files stored in your audit logs.
+	 */
+	// upload data
+	var files []plugintypes.AuditLogTransactionRequestFiles
+	al.Transaction_.Request_.Files_ = nil
+	for _, file := range tx.variables.files.Get("") {
+		var size int64
+		if fs := tx.variables.filesSizes.Get(file); len(fs) > 0 {
+			size, _ = strconv.ParseInt(fs[0], 10, 64)
+			// we ignore the error as it defaults to 0
+		}
+		ext := filepath.Ext(file)
+		at := auditlog.TransactionRequestFiles{
+			Size_: size,
+			Name_: file,
+			Mime_: mime.TypeByExtension(ext),
+		}
+		files = append(files, at)
+	}
+	al.Transaction_.Request_.Files_ = files
+}
+
+func (tx *Transaction) auditLogPartResponseBody(al *auditlog.Log) {
+	if al.Transaction_.Response_ == nil {
+		al.Transaction_.Response_ = &auditlog.TransactionResponse{}
+	}
+	al.Transaction_.Response_.Body_ = tx.variables.responseBody.Get()
+}
+
+func (tx *Transaction) auditLogPartResponseHeaders(al *auditlog.Log) {
+	if al.Transaction_.Response_ == nil {
+		al.Transaction_.Response_ = &auditlog.TransactionResponse{}
+	}
+	responseStatus, err := strconv.Atoi(tx.variables.responseStatus.Get())
+	if err != nil {
+		tx.debugLogger.Debug().Err(err).Msg("Error converting response status to integer")
+	}
+	al.Transaction_.Response_.Status_ = responseStatus
+	al.Transaction_.Response_.Headers_ = tx.variables.responseHeaders.Data()
+}
+
+func (tx *Transaction) auditLogPartProducer(al *auditlog.Log) {
+	al.Transaction_.Producer_ = &auditlog.TransactionProducer{
+		Connector_:  tx.WAF.ProducerConnector,
+		Version_:    tx.WAF.ProducerConnectorVersion,
+		Server_:     "",
+		RuleEngine_: tx.RuleEngine.String(),
+		Stopwatch_:  tx.GetStopWatch(),
+		Rulesets_:   tx.WAF.ComponentNames,
+	}
+}
+
+func (tx *Transaction) auditLogPartMatchedRules(al *auditlog.Log, includeErrorLog bool) {
+	for _, matchedRule := range tx.matchedRules {
+		// Log action is required to log a matched rule on both error log and audit log
+		// An assertion has to be done to check if the MatchedRule implements the Log() function before calling Log()
+		// It is performed to avoid breaking the Coraza v3.* API adding a Log() method to the MatchedRule interface
+		loggableRule, ok := matchedRule.(*corazarules.MatchedRule)
+		if ok && loggableRule.Log() {
+			r := matchedRule.Rule()
+			for _, matchData := range matchedRule.MatchedDatas() {
+				newAlEntry := auditlog.Message{
+					Actionset_: strings.Join(tx.WAF.ComponentNames, " "),
+					Message_:   matchData.Message(),
+					Data_: &auditlog.MessageData{
+						File_:     matchedRule.Rule().File(),
+						Line_:     matchedRule.Rule().Line(),
+						ID_:       r.ID(),
+						Rev_:      r.Revision(),
+						Msg_:      matchData.Message(),
+						Data_:     matchData.Data(),
+						Severity_: r.Severity(),
+						Ver_:      r.Version(),
+						Maturity_: r.Maturity(),
+						Accuracy_: r.Accuracy(),
+						Tags_:     r.Tags(),
+						Raw_:      r.Raw(),
+					},
+				}
+				// If AuditLogPartAuditLogTrailer (H) is set, we expect to log the error messages emitted by the rules
+				// in the audit log
+				if includeErrorLog {
+					newAlEntry.ErrorMessage_ = matchedRule.ErrorLog()
+				}
+				al.Messages_ = append(al.Messages_, newAlEntry)
+			}
+		}
+	}
 }
 
 // Close closes the transaction after phase 5
@@ -1653,785 +1689,3 @@ func (tx *Transaction) setTimeVariables() {
 	tx.variables.timeWday.Set(strconv.Itoa(int(timestamp.Weekday())))
 }
 
-// TransactionVariables has pointers to all the variables of the transaction
-type TransactionVariables struct {
-	args                     *collections.ConcatKeyed
-	argsCombinedSize         *collections.SizeCollection
-	argsGet                  *collections.NamedCollection
-	argsGetNames             collection.Keyed
-	argsNames                *collections.ConcatKeyed
-	argsPath                 *collections.NamedCollection
-	argsPost                 *collections.NamedCollection
-	argsPostNames            collection.Keyed
-	duration                 *collections.Single
-	env                      *collections.Map
-	files                    *collections.Map
-	filesCombinedSize        *collections.Single
-	filesNames               *collections.Map
-	filesSizes               *collections.Map
-	filesTmpContent          *collections.Map
-	filesTmpNames            *collections.Map
-	fullRequestLength        *collections.Single
-	geo                      *collections.Map
-	highestSeverity          *collections.Single
-	inboundDataError         *collections.Single
-	matchedVar               *collections.Single
-	matchedVarName           *collections.Single
-	matchedVars              *collections.NamedCollection
-	matchedVarsNames         collection.Keyed
-	multipartDataAfter       *collections.Single
-	multipartFilename        *collections.Map
-	multipartName            *collections.Map
-	multipartPartHeaders     *collections.Map
-	multipartStrictError     *collections.Single
-	outboundDataError        *collections.Single
-	queryString              *collections.Single
-	remoteAddr               *collections.Single
-	remoteHost               *collections.Single
-	remotePort               *collections.Single
-	reqbodyError             *collections.Single
-	reqbodyErrorMsg          *collections.Single
-	reqbodyProcessor         *collections.Single
-	reqbodyProcessorError    *collections.Single
-	reqbodyProcessorErrorMsg *collections.Single
-	requestBasename          *collections.Single
-	requestBody              *collections.Single
-	requestBodyLength        *collections.Single
-	requestCookies           *collections.NamedCollection
-	requestCookiesNames      collection.Keyed
-	requestFilename          *collections.Single
-	requestHeaders           *collections.NamedCollection
-	requestHeadersNames      collection.Keyed
-	requestLine              *collections.Single
-	requestMethod            *collections.Single
-	requestProtocol          *collections.Single
-	requestURI               *collections.Single
-	requestURIRaw            *collections.Single
-	requestXML               *collections.Map
-	responseBody             *collections.Single
-	responseContentLength    *collections.Single
-	responseContentType      *collections.Single
-	responseHeaders          *collections.NamedCollection
-	responseHeadersNames     collection.Keyed
-	responseProtocol         *collections.Single
-	responseStatus           *collections.Single
-	responseXML              *collections.Map
-	responseArgs             *collections.Map
-	resBodyProcessor         *collections.Single
-	rule                     *collections.Map
-	serverAddr               *collections.Single
-	serverName               *collections.Single
-	serverPort               *collections.Single
-	statusLine               *collections.Single
-	tx                       *collections.Map
-	uniqueID                 *collections.Single
-	urlencodedError          *collections.Single
-	xml                      *collections.Map
-	resBodyError             *collections.Single
-	resBodyErrorMsg          *collections.Single
-	resBodyProcessorError    *collections.Single
-	resBodyProcessorErrorMsg *collections.Single
-	time                     *collections.Single
-	timeDay                  *collections.Single
-	timeEpoch                *collections.Single
-	timeHour                 *collections.Single
-	timeMin                  *collections.Single
-	timeMon                  *collections.Single
-	timeSec                  *collections.Single
-	timeWday                 *collections.Single
-	timeYear                 *collections.Single
-}
-
-func NewTransactionVariables() *TransactionVariables {
-	v := &TransactionVariables{}
-	v.urlencodedError = collections.NewSingle(variables.UrlencodedError)
-	v.responseContentType = collections.NewSingle(variables.ResponseContentType)
-	v.uniqueID = collections.NewSingle(variables.UniqueID)
-	v.filesCombinedSize = collections.NewSingle(variables.FilesCombinedSize)
-	v.fullRequestLength = collections.NewSingle(variables.FullRequestLength)
-	v.inboundDataError = collections.NewSingle(variables.InboundDataError)
-	v.matchedVar = collections.NewSingle(variables.MatchedVar)
-	v.matchedVarName = collections.NewSingle(variables.MatchedVarName)
-	v.multipartDataAfter = collections.NewSingle(variables.MultipartDataAfter)
-	v.outboundDataError = collections.NewSingle(variables.OutboundDataError)
-	v.queryString = collections.NewSingle(variables.QueryString)
-	v.remoteAddr = collections.NewSingle(variables.RemoteAddr)
-	v.remoteHost = collections.NewSingle(variables.RemoteHost)
-	v.remotePort = collections.NewSingle(variables.RemotePort)
-	v.reqbodyError = collections.NewSingle(variables.ReqbodyError)
-	v.reqbodyErrorMsg = collections.NewSingle(variables.ReqbodyErrorMsg)
-	v.reqbodyProcessorError = collections.NewSingle(variables.ReqbodyProcessorError)
-	v.reqbodyProcessorErrorMsg = collections.NewSingle(variables.ReqbodyProcessorErrorMsg)
-	v.reqbodyProcessor = collections.NewSingle(variables.ReqbodyProcessor)
-	v.requestBasename = collections.NewSingle(variables.RequestBasename)
-	v.requestBody = collections.NewSingle(variables.RequestBody)
-	v.requestBodyLength = collections.NewSingle(variables.RequestBodyLength)
-	v.requestFilename = collections.NewSingle(variables.RequestFilename)
-	v.requestLine = collections.NewSingle(variables.RequestLine)
-	v.requestMethod = collections.NewSingle(variables.RequestMethod)
-	v.requestProtocol = collections.NewSingle(variables.RequestProtocol)
-	v.requestURI = collections.NewSingle(variables.RequestURI)
-	v.requestURIRaw = collections.NewSingle(variables.RequestURIRaw)
-	v.responseBody = collections.NewSingle(variables.ResponseBody)
-	v.responseContentLength = collections.NewSingle(variables.ResponseContentLength)
-	v.responseProtocol = collections.NewSingle(variables.ResponseProtocol)
-	v.responseStatus = collections.NewSingle(variables.ResponseStatus)
-	v.responseArgs = collections.NewMap(variables.ResponseArgs)
-	v.serverAddr = collections.NewSingle(variables.ServerAddr)
-	v.serverName = collections.NewSingle(variables.ServerName)
-	v.serverPort = collections.NewSingle(variables.ServerPort)
-	v.highestSeverity = collections.NewSingle(variables.HighestSeverity)
-	v.statusLine = collections.NewSingle(variables.StatusLine)
-	v.duration = collections.NewSingle(variables.Duration)
-	v.resBodyError = collections.NewSingle(variables.ResBodyError)
-	v.resBodyErrorMsg = collections.NewSingle(variables.ResBodyErrorMsg)
-	v.resBodyProcessorError = collections.NewSingle(variables.ResBodyProcessorError)
-	v.resBodyProcessorErrorMsg = collections.NewSingle(variables.ResBodyProcessorErrorMsg)
-
-	v.filesSizes = collections.NewMap(variables.FilesSizes)
-	v.filesTmpContent = collections.NewMap(variables.FilesTmpContent)
-	v.multipartFilename = collections.NewMap(variables.MultipartFilename)
-	v.multipartName = collections.NewMap(variables.MultipartName)
-	v.matchedVars = collections.NewNamedCollection(variables.MatchedVars)
-	v.matchedVarsNames = v.matchedVars.Names(variables.MatchedVarsNames)
-	v.requestCookies = collections.NewNamedCollection(variables.RequestCookies)
-	v.requestCookiesNames = v.requestCookies.Names(variables.RequestCookiesNames)
-	v.requestHeaders = collections.NewNamedCollection(variables.RequestHeaders)
-	v.requestHeadersNames = v.requestHeaders.Names(variables.RequestHeadersNames)
-	v.responseHeaders = collections.NewNamedCollection(variables.ResponseHeaders)
-	v.responseHeadersNames = v.responseHeaders.Names(variables.ResponseHeadersNames)
-	v.resBodyProcessor = collections.NewSingle(variables.ResBodyProcessor)
-	v.geo = collections.NewMap(variables.Geo)
-	v.tx = collections.NewMap(variables.TX)
-	v.rule = collections.NewMap(variables.Rule)
-	v.env = collections.NewMap(variables.Env)
-	v.files = collections.NewMap(variables.Files)
-	v.filesNames = collections.NewMap(variables.FilesNames)
-	v.filesTmpNames = collections.NewMap(variables.FilesTmpNames)
-	v.responseXML = collections.NewMap(variables.ResponseXML)
-	v.requestXML = collections.NewMap(variables.RequestXML)
-	v.multipartPartHeaders = collections.NewMap(variables.MultipartPartHeaders)
-	v.multipartStrictError = collections.NewSingle(variables.MultipartStrictError)
-	v.time = collections.NewSingle(variables.Time)
-	v.timeDay = collections.NewSingle(variables.TimeDay)
-	v.timeEpoch = collections.NewSingle(variables.TimeEpoch)
-	v.timeHour = collections.NewSingle(variables.TimeHour)
-	v.timeMin = collections.NewSingle(variables.TimeMin)
-	v.timeMon = collections.NewSingle(variables.TimeMon)
-	v.timeSec = collections.NewSingle(variables.TimeSec)
-	v.timeWday = collections.NewSingle(variables.TimeWday)
-	v.timeYear = collections.NewSingle(variables.TimeYear)
-
-	// XML is a pointer to RequestXML
-	v.xml = v.requestXML
-
-	if shouldUseCaseSensitiveNamedCollection {
-		v.argsGet = collections.NewCaseSensitiveNamedCollection(variables.ArgsGet)
-		v.argsPost = collections.NewCaseSensitiveNamedCollection(variables.ArgsPost)
-		v.argsPath = collections.NewCaseSensitiveNamedCollection(variables.ArgsPath)
-	} else {
-		v.argsGet = collections.NewNamedCollection(variables.ArgsGet)
-		v.argsPost = collections.NewNamedCollection(variables.ArgsPost)
-		v.argsPath = collections.NewNamedCollection(variables.ArgsPath)
-	}
-
-	v.argsGetNames = v.argsGet.Names(variables.ArgsGetNames)
-	v.argsPostNames = v.argsPost.Names(variables.ArgsPostNames)
-	v.argsCombinedSize = collections.NewSizeCollection(variables.ArgsCombinedSize, v.argsGet, v.argsPost)
-	v.args = collections.NewConcatKeyed(
-		variables.Args,
-		v.argsGet,
-		v.argsPost,
-		v.argsPath,
-	)
-	v.argsNames = collections.NewConcatKeyed(
-		variables.ArgsNames,
-		v.argsGetNames,
-		v.argsPostNames,
-		// Only used in a concatenating collection so variable name doesn't matter.
-		v.argsPath.Names(variables.Unknown),
-	)
-	return v
-}
-
-func (v *TransactionVariables) UrlencodedError() collection.Single {
-	return v.urlencodedError
-}
-
-func (v *TransactionVariables) ResponseContentType() collection.Single {
-	return v.responseContentType
-}
-
-func (v *TransactionVariables) UniqueID() collection.Single {
-	return v.uniqueID
-}
-
-func (v *TransactionVariables) ArgsCombinedSize() collection.Collection {
-	return v.argsCombinedSize
-}
-
-func (v *TransactionVariables) FilesCombinedSize() collection.Single {
-	return v.filesCombinedSize
-}
-
-func (v *TransactionVariables) FullRequestLength() collection.Single {
-	return v.fullRequestLength
-}
-
-func (v *TransactionVariables) InboundDataError() collection.Single {
-	return v.inboundDataError
-}
-
-func (v *TransactionVariables) MatchedVar() collection.Single {
-	return v.matchedVar
-}
-
-func (v *TransactionVariables) MatchedVarName() collection.Single {
-	return v.matchedVarName
-}
-
-func (v *TransactionVariables) MultipartDataAfter() collection.Single {
-	return v.multipartDataAfter
-}
-
-func (v *TransactionVariables) MultipartPartHeaders() collection.Map {
-	return v.multipartPartHeaders
-}
-
-func (v *TransactionVariables) OutboundDataError() collection.Single {
-	return v.outboundDataError
-}
-
-func (v *TransactionVariables) QueryString() collection.Single {
-	return v.queryString
-}
-
-func (v *TransactionVariables) RemoteAddr() collection.Single {
-	return v.remoteAddr
-}
-
-func (v *TransactionVariables) RemoteHost() collection.Single {
-	return v.remoteHost
-}
-
-func (v *TransactionVariables) RemotePort() collection.Single {
-	return v.remotePort
-}
-
-func (v *TransactionVariables) RequestBodyError() collection.Single {
-	return v.reqbodyError
-}
-
-func (v *TransactionVariables) RequestBodyErrorMsg() collection.Single {
-	return v.reqbodyErrorMsg
-}
-
-func (v *TransactionVariables) RequestBodyProcessorError() collection.Single {
-	return v.reqbodyProcessorError
-}
-
-func (v *TransactionVariables) RequestBodyProcessorErrorMsg() collection.Single {
-	return v.reqbodyProcessorErrorMsg
-}
-
-func (v *TransactionVariables) RequestBodyProcessor() collection.Single {
-	return v.reqbodyProcessor
-}
-
-func (v *TransactionVariables) RequestBasename() collection.Single {
-	return v.requestBasename
-}
-
-func (v *TransactionVariables) RequestBody() collection.Single {
-	return v.requestBody
-}
-
-func (v *TransactionVariables) RequestBodyLength() collection.Single {
-	return v.requestBodyLength
-}
-
-func (v *TransactionVariables) RequestFilename() collection.Single {
-	return v.requestFilename
-}
-
-func (v *TransactionVariables) RequestLine() collection.Single {
-	return v.requestLine
-}
-
-func (v *TransactionVariables) RequestMethod() collection.Single {
-	return v.requestMethod
-}
-
-func (v *TransactionVariables) RequestProtocol() collection.Single {
-	return v.requestProtocol
-}
-
-func (v *TransactionVariables) RequestURI() collection.Single {
-	return v.requestURI
-}
-
-func (v *TransactionVariables) RequestURIRaw() collection.Single {
-	return v.requestURIRaw
-}
-
-func (v *TransactionVariables) ResponseBody() collection.Single {
-	return v.responseBody
-}
-
-func (v *TransactionVariables) ResponseContentLength() collection.Single {
-	return v.responseContentLength
-}
-
-func (v *TransactionVariables) ResponseProtocol() collection.Single {
-	return v.responseProtocol
-}
-
-func (v *TransactionVariables) ResponseStatus() collection.Single {
-	return v.responseStatus
-}
-
-func (v *TransactionVariables) ServerAddr() collection.Single {
-	return v.serverAddr
-}
-
-func (v *TransactionVariables) ServerName() collection.Single {
-	return v.serverName
-}
-
-func (v *TransactionVariables) ServerPort() collection.Single {
-	return v.serverPort
-}
-
-func (v *TransactionVariables) HighestSeverity() collection.Single {
-	return v.highestSeverity
-}
-
-func (v *TransactionVariables) StatusLine() collection.Single {
-	return v.statusLine
-}
-
-func (v *TransactionVariables) Env() collection.Map {
-	return v.env
-}
-
-func (v *TransactionVariables) TX() collection.Map {
-	return v.tx
-}
-
-func (v *TransactionVariables) Rule() collection.Map {
-	return v.rule
-}
-
-func (v *TransactionVariables) Duration() collection.Single {
-	return v.duration
-}
-
-func (v *TransactionVariables) Args() collection.Keyed {
-	return v.args
-}
-
-func (v *TransactionVariables) ArgsGet() collection.Map {
-	return v.argsGet
-}
-
-func (v *TransactionVariables) ArgsPost() collection.Map {
-	return v.argsPost
-}
-
-func (v *TransactionVariables) ArgsPath() collection.Map {
-	return v.argsPath
-}
-
-func (v *TransactionVariables) FilesTmpNames() collection.Map {
-	return v.filesTmpNames
-}
-
-func (v *TransactionVariables) Geo() collection.Map {
-	return v.geo
-}
-
-func (v *TransactionVariables) Files() collection.Map {
-	return v.files
-}
-
-func (v *TransactionVariables) RequestCookies() collection.Map {
-	return v.requestCookies
-}
-
-func (v *TransactionVariables) RequestHeaders() collection.Map {
-	return v.requestHeaders
-}
-
-func (v *TransactionVariables) ResponseHeaders() collection.Map {
-	return v.responseHeaders
-}
-
-func (v *TransactionVariables) MultipartName() collection.Map {
-	return v.multipartName
-}
-
-func (v *TransactionVariables) MatchedVarsNames() collection.Keyed {
-	return v.matchedVarsNames
-}
-
-func (v *TransactionVariables) MultipartFilename() collection.Map {
-	return v.multipartFilename
-}
-
-func (v *TransactionVariables) MatchedVars() collection.Map {
-	return v.matchedVars
-}
-
-func (v *TransactionVariables) FilesSizes() collection.Map {
-	return v.filesSizes
-}
-
-func (v *TransactionVariables) FilesNames() collection.Map {
-	return v.filesNames
-}
-
-func (v *TransactionVariables) FilesTmpContent() collection.Map {
-	return v.filesTmpContent
-}
-
-func (v *TransactionVariables) ResponseHeadersNames() collection.Keyed {
-	return v.responseHeadersNames
-}
-
-func (v *TransactionVariables) ResponseArgs() collection.Map {
-	return v.responseArgs
-}
-
-func (v *TransactionVariables) RequestHeadersNames() collection.Keyed {
-	return v.requestHeadersNames
-}
-
-func (v *TransactionVariables) RequestCookiesNames() collection.Keyed {
-	return v.requestCookiesNames
-}
-
-func (v *TransactionVariables) XML() collection.Map {
-	return v.xml
-}
-
-func (v *TransactionVariables) RequestXML() collection.Map {
-	return v.requestXML
-}
-
-func (v *TransactionVariables) ResponseXML() collection.Map {
-	return v.responseXML
-}
-
-func (v *TransactionVariables) ResponseBodyProcessor() collection.Single {
-	return v.resBodyProcessor
-}
-
-func (v *TransactionVariables) ArgsNames() collection.Keyed {
-	return v.argsNames
-}
-
-func (v *TransactionVariables) ArgsGetNames() collection.Keyed {
-	return v.argsGetNames
-}
-
-func (v *TransactionVariables) ArgsPostNames() collection.Keyed {
-	return v.argsPostNames
-}
-
-func (v *TransactionVariables) ResBodyError() collection.Single {
-	return v.resBodyError
-}
-
-func (v *TransactionVariables) ResBodyErrorMsg() collection.Single {
-	return v.resBodyErrorMsg
-}
-
-func (v *TransactionVariables) ResBodyProcessorError() collection.Single {
-	return v.resBodyProcessorError
-}
-
-func (v *TransactionVariables) ResBodyProcessorErrorMsg() collection.Single {
-	return v.resBodyProcessorErrorMsg
-}
-
-func (v *TransactionVariables) MultipartStrictError() collection.Single {
-	return v.multipartStrictError
-}
-
-// All iterates over the variables. We return both variable and its collection, i.e. key/value, to follow
-// general range iteration in Go which always has a key and value (key is int index for slices). Notably,
-// this is consistent with discussions for custom iterable types in a future language version
-// https://github.com/golang/go/discussions/56413
-func (v *TransactionVariables) All(f func(v variables.RuleVariable, col collection.Collection) bool) {
-	if !f(variables.Args, v.args) {
-		return
-	}
-	if !f(variables.ArgsCombinedSize, v.argsCombinedSize) {
-		return
-	}
-	if !f(variables.ArgsGet, v.argsGet) {
-		return
-	}
-	if !f(variables.ArgsGetNames, v.argsGetNames) {
-		return
-	}
-	if !f(variables.ArgsNames, v.argsNames) {
-		return
-	}
-	if !f(variables.ArgsPath, v.argsPath) {
-		return
-	}
-	if !f(variables.ArgsPost, v.argsPost) {
-		return
-	}
-	if !f(variables.ArgsPostNames, v.argsPostNames) {
-		return
-	}
-	if !f(variables.Duration, v.duration) {
-		return
-	}
-	if !f(variables.Env, v.env) {
-		return
-	}
-	if !f(variables.Files, v.files) {
-		return
-	}
-	if !f(variables.FilesCombinedSize, v.filesCombinedSize) {
-		return
-	}
-	if !f(variables.FilesNames, v.filesNames) {
-		return
-	}
-	if !f(variables.FilesSizes, v.filesSizes) {
-		return
-	}
-	if !f(variables.FilesTmpContent, v.filesTmpContent) {
-		return
-	}
-	if !f(variables.FilesTmpNames, v.filesTmpNames) {
-		return
-	}
-	if !f(variables.FullRequestLength, v.fullRequestLength) {
-		return
-	}
-	if !f(variables.Geo, v.geo) {
-		return
-	}
-	if !f(variables.HighestSeverity, v.highestSeverity) {
-		return
-	}
-	if !f(variables.InboundDataError, v.inboundDataError) {
-		return
-	}
-	if !f(variables.MatchedVar, v.matchedVar) {
-		return
-	}
-	if !f(variables.MatchedVarName, v.matchedVarName) {
-		return
-	}
-	if !f(variables.MatchedVars, v.matchedVars) {
-		return
-	}
-	if !f(variables.MatchedVarsNames, v.matchedVarsNames) {
-		return
-	}
-	if !f(variables.MultipartDataAfter, v.multipartDataAfter) {
-		return
-	}
-	if !f(variables.MultipartFilename, v.multipartFilename) {
-		return
-	}
-	if !f(variables.MultipartName, v.multipartName) {
-		return
-	}
-	if !f(variables.MultipartPartHeaders, v.multipartPartHeaders) {
-		return
-	}
-	if !f(variables.MultipartStrictError, v.multipartStrictError) {
-		return
-	}
-	if !f(variables.OutboundDataError, v.outboundDataError) {
-		return
-	}
-	if !f(variables.QueryString, v.queryString) {
-		return
-	}
-	if !f(variables.RemoteAddr, v.remoteAddr) {
-		return
-	}
-	if !f(variables.RemoteHost, v.remoteHost) {
-		return
-	}
-	if !f(variables.RemotePort, v.remotePort) {
-		return
-	}
-	if !f(variables.ReqbodyError, v.reqbodyError) {
-		return
-	}
-	if !f(variables.ReqbodyErrorMsg, v.reqbodyErrorMsg) {
-		return
-	}
-	if !f(variables.ReqbodyProcessor, v.reqbodyProcessor) {
-		return
-	}
-	if !f(variables.ReqbodyProcessorError, v.reqbodyProcessorError) {
-		return
-	}
-	if !f(variables.ReqbodyProcessorErrorMsg, v.reqbodyProcessorErrorMsg) {
-		return
-	}
-	if !f(variables.RequestBasename, v.requestBasename) {
-		return
-	}
-	if !f(variables.RequestBody, v.requestBody) {
-		return
-	}
-	if !f(variables.RequestBodyLength, v.requestBodyLength) {
-		return
-	}
-	if !f(variables.RequestCookies, v.requestCookies) {
-		return
-	}
-	if !f(variables.RequestCookiesNames, v.requestCookiesNames) {
-		return
-	}
-	if !f(variables.RequestFilename, v.requestFilename) {
-		return
-	}
-	if !f(variables.RequestHeaders, v.requestHeaders) {
-		return
-	}
-	if !f(variables.RequestHeadersNames, v.requestHeadersNames) {
-		return
-	}
-	if !f(variables.RequestLine, v.requestLine) {
-		return
-	}
-	if !f(variables.RequestMethod, v.requestMethod) {
-		return
-	}
-	if !f(variables.RequestProtocol, v.requestProtocol) {
-		return
-	}
-	if !f(variables.RequestURI, v.requestURI) {
-		return
-	}
-	if !f(variables.RequestURIRaw, v.requestURIRaw) {
-		return
-	}
-	if !f(variables.RequestXML, v.requestXML) {
-		return
-	}
-	if !f(variables.ResponseBody, v.responseBody) {
-		return
-	}
-	if !f(variables.ResponseContentLength, v.responseContentLength) {
-		return
-	}
-	if !f(variables.ResponseContentType, v.responseContentType) {
-		return
-	}
-	if !f(variables.ResponseHeaders, v.responseHeaders) {
-		return
-	}
-	if !f(variables.ResponseHeadersNames, v.responseHeadersNames) {
-		return
-	}
-	if !f(variables.ResponseProtocol, v.responseProtocol) {
-		return
-	}
-	if !f(variables.ResponseStatus, v.responseStatus) {
-		return
-	}
-	if !f(variables.ResponseXML, v.responseXML) {
-		return
-	}
-	if !f(variables.ResponseArgs, v.responseArgs) {
-		return
-	}
-	if !f(variables.ResBodyProcessor, v.resBodyProcessor) {
-		return
-	}
-	if !f(variables.Rule, v.rule) {
-		return
-	}
-	if !f(variables.ServerAddr, v.serverAddr) {
-		return
-	}
-	if !f(variables.ServerName, v.serverName) {
-		return
-	}
-	if !f(variables.ServerPort, v.serverPort) {
-		return
-	}
-	if !f(variables.StatusLine, v.statusLine) {
-		return
-	}
-	if !f(variables.TX, v.tx) {
-		return
-	}
-	if !f(variables.UniqueID, v.uniqueID) {
-		return
-	}
-	if !f(variables.UrlencodedError, v.urlencodedError) {
-		return
-	}
-	if !f(variables.XML, v.xml) {
-		return
-	}
-	if !f(variables.Time, v.time) {
-		return
-	}
-	if !f(variables.TimeDay, v.timeDay) {
-		return
-	}
-	if !f(variables.TimeEpoch, v.timeEpoch) {
-		return
-	}
-	if !f(variables.TimeHour, v.timeHour) {
-		return
-	}
-	if !f(variables.TimeMin, v.timeMin) {
-		return
-	}
-	if !f(variables.TimeMon, v.timeMon) {
-		return
-	}
-	if !f(variables.TimeSec, v.timeSec) {
-		return
-	}
-	if !f(variables.TimeWday, v.timeWday) {
-		return
-	}
-	if !f(variables.TimeYear, v.timeYear) {
-		return
-	}
-}
-
-type formattable interface {
-	Format(res *strings.Builder)
-}
-
-func (v *TransactionVariables) format(res *strings.Builder) {
-	v.All(func(_ variables.RuleVariable, col collection.Collection) bool {
-		if f, ok := col.(formattable); ok {
-			f.Format(res)
-		} else {
-			fmt.Fprintln(res, col)
-		}
-		return true
-	})
-}
-
-type resettable interface {
-	Reset()
-}
-
-func (v *TransactionVariables) reset() {
-	v.All(func(_ variables.RuleVariable, col collection.Collection) bool {
-		if r, ok := col.(resettable); ok {
-			r.Reset()
-		}
-		return true
-	})
-}
